remaining task: authentication and authorization move views to REST API document api

+++ src/main/java/com/venturegardengroup/foodvendorapplication/repositories/MessageStatusRepository.java
@@ -3,6 +3,6 @@ package com.venturegardengroup.foodvendorapplication.repositories;
 import com.venturegardengroup.foodvendorapplication.models.MessageStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageStatusRepository extends JpaRepository<MessageStatus,Long> {
+public interface MessageStatusRepository extends JpaRepository<MessageStatus,Integer> {
 
 }
