+++ src/main/java/com/venturegardengroup/foodvendorapplication/repositories/OrderStatusRepository.java
@@ -3,6 +3,6 @@ package com.venturegardengroup.foodvendorapplication.repositories;
 import com.venturegardengroup.foodvendorapplication.models.OrderStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderStatusRepository extends JpaRepository<OrderStatus,Long> {
+public interface OrderStatusRepository extends JpaRepository<OrderStatus,Integer> {
 
 }
